Move param display code from helpers into parm_displayers, use param_formatters.parse_value
